@@ -5,13 +5,14 @@
   "description": "stream-proxy",
   "directories": {},
   "scripts": {},
+  "type": "module",
   "dependencies": {
-    "express": "^4.17.1",
-    "request": "^2.88.2"
+    "cors": "^2.8.5",
+    "express": "^4.17.2",
+    "got": "^12.0.0"
   },
-  "devDependencies": {},
   "engines": {
-    "node": ">= 12.0.0",
+    "node": ">= 14.0.0",
     "npm": ">= 6.0.0"
   }
 }
